Move arithmetic and comparison functions out of one-to-one matching operator file
